Moved SelectionHighlightController.
The SelectionHighlightController is a useful class that can be used in
combination with the ScreenSelector class. Thus, changed the
ScreenSelection example application by pulling out the
SelectionHighlightController and moving it to the
gov.nasa.worldwindx.examples.util package.
